command adding
obejctActive,  createPrefeb, distroyPrefab, function added, move rotate function modified, (move or rotate by adding parameter when function called)

obejctActive,  createPrefeb, distroyPrefa 기능 추가, move, rotate함수 수정(함수가 호출되었을 때의 인자를 더하는 방식으로 )
